requirements.txt
line-bot-sdk

flask

pymysql
os